+++ services/gateway/handler/auth/service_test.go
@@ -14,6 +14,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	macaroon "gopkg.in/macaroon.v2"
 	mc "github.com/idlab-discover/kebeng/services/gateway/internal/macaroon"
+	"github.com/idlab-discover/kebeng/services/gateway/internal/model"
 )
 
 // validMacaroonForTest is a helper that deserializes a macaroon string using our MacaroonDeserialize.
@@ -30,7 +31,7 @@ func validMacaroonForTest(t *testing.T, m *macaroon.Macaroon) string {
 // TestGenerateMacaroon_Success tests a successful generation.
 func TestGenerateMacaroon_Success(t *testing.T) {
 	ctx := context.Background()
-	req := &GenerateMacaroonRequest{
+	req := &model.GenerateMacaroonRequest{
 		Permissions: []string{"edit_account", "modify_account_key"},
 		Channels:    []string{"stable", "beta"},
 		Expires:     "", // let expiry be computed automatically if a default permission is found
@@ -81,7 +82,7 @@ func TestGenerateMacaroon_Success(t *testing.T) {
 func TestGenerateMacaroon_Failure(t *testing.T) {
 	ctx := context.Background()
 	// lets pass in a bad expires field
-	req := &GenerateMacaroonRequest{
+	req := &model.GenerateMacaroonRequest{
 		Permissions: []string{"edit_account"},
 		Channels:    []string{"stable"},
 		Expires:     "bad-expires", // invalid expiry
@@ -104,15 +105,15 @@ func TestGenerateMacaroon_Failure(t *testing.T) {
 func TestValidateGenerateMacaroonRequest(t *testing.T) {
 	tests := []struct {
 		name         string
-		req          *GenerateMacaroonRequest
+		req          *model.GenerateMacaroonRequest
 		expectErrors []string
 	}{
 		{
 			name: "valid request",
-			req: &GenerateMacaroonRequest{
+			req: &model.GenerateMacaroonRequest{
 				Permissions: []string{"edit_account", "package_access"},
 				Channels:    []string{"stable", "edge"},
-				Packages: []PackageRestriction{
+				Packages: []model.PackageRestriction{
 					// Valid package using snap_id only.
 					{SnapId: "snap1"},
 					// Valid package using name and series.
@@ -124,30 +125,30 @@ func TestValidateGenerateMacaroonRequest(t *testing.T) {
 		},
 		{
 			name: "invalid permission",
-			req: &GenerateMacaroonRequest{
+			req: &model.GenerateMacaroonRequest{
 				Permissions: []string{"edit_account", "invalid_perm"},
 				Channels:    []string{"stable"},
-				Packages:    []PackageRestriction{},
+				Packages:    []model.PackageRestriction{},
 				Expires:     "",
 			},
 			expectErrors: []string{"permission value 'invalid_perm' is not allowed"},
 		},
 		{
 			name: "invalid channel",
-			req: &GenerateMacaroonRequest{
+			req: &model.GenerateMacaroonRequest{
 				Permissions: []string{"edit_account"},
 				Channels:    []string{"invalid_channel"},
-				Packages:    []PackageRestriction{},
+				Packages:    []model.PackageRestriction{},
 				Expires:     "",
 			},
 			expectErrors: []string{"channel value 'invalid_channel' is not allowed"},
 		},
 		{
 			name: "package with nothing provided",
-			req: &GenerateMacaroonRequest{
+			req: &model.GenerateMacaroonRequest{
 				Permissions: []string{"edit_account"},
 				Channels:    []string{"stable"},
-				Packages: []PackageRestriction{
+				Packages: []model.PackageRestriction{
 					{Name: "", Series: "", SnapId: ""},
 				},
 				Expires: "",
@@ -156,10 +157,10 @@ func TestValidateGenerateMacaroonRequest(t *testing.T) {
 		},
 		{
 			name: "package with snap_id and extra fields",
-			req: &GenerateMacaroonRequest{
+			req: &model.GenerateMacaroonRequest{
 				Permissions: []string{"edit_account"},
 				Channels:    []string{"stable"},
-				Packages: []PackageRestriction{
+				Packages: []model.PackageRestriction{
 					{Name: "pkg1", Series: "16", SnapId: "snap1"},
 				},
 				Expires: "",
@@ -168,10 +169,10 @@ func TestValidateGenerateMacaroonRequest(t *testing.T) {
 		},
 		{
 			name: "package missing name or series",
-			req: &GenerateMacaroonRequest{
+			req: &model.GenerateMacaroonRequest{
 				Permissions: []string{"edit_account"},
 				Channels:    []string{"stable"},
-				Packages: []PackageRestriction{
+				Packages: []model.PackageRestriction{
 					{Name: "pkg1", Series: "", SnapId: ""},
 					{Name: "", Series: "16", SnapId: ""},
 				},
@@ -184,10 +185,10 @@ func TestValidateGenerateMacaroonRequest(t *testing.T) {
 		},
 		{
 			name: "invalid expires format",
-			req: &GenerateMacaroonRequest{
+			req: &model.GenerateMacaroonRequest{
 				Permissions: []string{"edit_account"},
 				Channels:    []string{"stable"},
-				Packages:    []PackageRestriction{},
+				Packages:    []model.PackageRestriction{},
 				Expires:     "invalid-date",
 			},
 			expectErrors: []string{"expires:"}, // we look for the substring "expires:" in the error
