+++ services/gateway/internal/model/auth.go
@@ -1,4 +1,4 @@
-package auth
+package model
 
 import "github.com/idlab-discover/kebeng/common/cerror"
